Moving function prototypes from lin-city.h
Moving the function prototypes from lin-city.h to the header file where
the function definition belongs, and including those headers in the
sources that use the function.

If the function prototype is not implemented then erase it.

Add includes to the new header file where the function is used.

Moving the functions in shrtypes.c to engine.c, where they belongs.