Merge separate coordinate variables with a struct
In englobs.h some info that store coordinates had separate variables for
the x coordinate and the y coordinates (for example: mappoint_array_x[]
and mappoint_array_y[]). With the use of a new type Map_Coord to hold
both coordinates we can merge each of these variables (or arrays) into
one variable. This has been done in the case of mappoint_array[],
last_built, markets[] and substations[].

Also, the variables numof_markets and numof_substations, that stores the
number of elements in their related arrays, has been changed its type
from int to unsigned int to reflect the fact.